Adding progressbar feature
While it is not straitfoward (added an ugly "see progress button" in requests template, the main work is done to make P2N able to communicate with the front end using SSE connexions using a message announcer and a listener of get requests.
Some extra design is also needed. Please FIX IT.